refactor(core-model): introduce shared base class for component hierarchy
