--- client/src/app/core-model-admin/core-model-admin.component.scss
@@ -1,54 +1,0 @@
-// Table Wrapper Styles
-.table-wrapper {
-  align-items: center;
-  border-radius: 12px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.5);
-  display: flex;
-  flex-direction: column;
-  padding: 20px;
-  max-width: 100%;
-}
-
-// Import and Extend Container Styles
-.import-extend-download-container {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  align-items: center;
-  gap: 20px;
-}
-
-// Filter and Pagination Container Styles
-.filter-pagination-container {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  align-items: center;
-  border-bottom: 3px solid #e0e0e0;
-}
-
-.table-container {
-  max-width: 95%;
-}
-
-// Filter Container Styles
-.filter-container {
-  mat-form-field {
-    width: 100%;
-    max-width: 300px;
-  }
-}
-
-// Pagination Container Styles
-.pagination-container {
-  justify-self: end;
-}
-
-.actions-buttons {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.button-style.first {
-  display: flex;
-  margin-bottom: 8px;
-}
