+++ client/src/app/core-model/base/core-model-base.scss
@@ -9,13 +9,6 @@
   max-width: 100%;
 }
 
-// Import and Extend Container Styles
-.download-button-container {
-  display: flex;
-  align-items: center;
-  gap: 20px;
-}
-
 // Filter and Pagination Container Styles
 .filter-pagination-container {
   display: grid;
