+++ client/src/app/core-model/core-model-admin/core-model-admin.component.html
