+++ client/src/app/core-model/core-model-admin/core-model-admin.component.scss
@@ -1,0 +1,9 @@
+@use "../base/core-model-base.scss";
+
+// Import and Extend Container Styles
+.import-extend-download-container {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  align-items: center;
+  gap: 20px;
+}
