+++ client/src/app/core-model/core-model/core-model.component.html
@@ -214,35 +214,52 @@
           </td>
         </ng-container>
 
-        <!-- Study1 Variable -->
-        <ng-container matColumnDef="study1Variable">
+        <!-- Dynamic Study Columns -->
+        @for (studyName of studyColumnNames; track studyName) {
+        <ng-container
+          [matColumnDef]="this.tableService.toCamelCase(studyName) + 'Variable'"
+        >
           <th mat-header-cell *matHeaderCellDef mat-sort-header>
-            Study1 Variable
+            {{ studyName }} Variable
           </th>
-          <td mat-cell *matCellDef="let element" class="cell-bg-study1">
-            {{ element.studies[0].variable }}
+          <td
+            mat-cell
+            *matCellDef="let element"
+            [class]="this.tableService.getStudyClass(studyName)"
+          >
+            {{
+              this.tableService.getStudyField(
+                element,
+                this.tableService.toCamelCase(studyName),
+                "variable"
+              )
+            }}
           </td>
         </ng-container>
-
-        <!-- Study1 Description -->
-        <ng-container matColumnDef="study1Description">
+        } @for (studyName of studyColumnNames; track studyName) {
+        <ng-container
+          [matColumnDef]="
+            this.tableService.toCamelCase(studyName) + 'Description'
+          "
+        >
           <th mat-header-cell *matHeaderCellDef mat-sort-header>
-            Study1 Description
+            {{ studyName }} Description
           </th>
-          <td mat-cell *matCellDef="let element" class="cell-bg-study1">
-            {{ element.studies[0].description }}
+          <td
+            mat-cell
+            *matCellDef="let element"
+            [class]="this.tableService.getStudyClass(studyName)"
+          >
+            {{
+              this.tableService.getStudyField(
+                element,
+                this.tableService.toCamelCase(studyName),
+                "description"
+              )
+            }}
           </td>
         </ng-container>
-
-        <!-- Study2 Variable -->
-        <ng-container matColumnDef="study2Variable">
-          <th mat-header-cell *matHeaderCellDef mat-sort-header>
-            Study2 Variable
-          </th>
-          <td mat-cell *matCellDef="let element" class="cell-bg-study2">
-            {{ element.studies[1].variable }}
-          </td>
-        </ng-container>
+        }
 
         <tr
           mat-header-row
