+++ client/src/app/core-model/core-model/core-model.component.spec.ts
