+++ client/src/app/core-model/core-model/core-model.component.ts
@@ -1,0 +1,43 @@
+import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatSort, MatSortModule } from '@angular/material/sort';
+import { MatTableModule } from '@angular/material/table';
+
+import { CoreModelBase } from '../base/core-model-base';
+
+@Component({
+  selector: 'app-core-model-table',
+  standalone: true,
+  imports: [
+    MatFormFieldModule,
+    MatIconModule,
+    MatInputModule,
+    MatPaginatorModule,
+    MatProgressSpinnerModule,
+    MatSortModule,
+    MatTableModule,
+  ],
+  templateUrl: './core-model.component.html',
+  styleUrl: './core-model.component.scss',
+})
+export class CoreModelComponent
+  extends CoreModelBase
+  implements OnInit, OnDestroy
+{
+  @ViewChild(MatPaginator) paginator!: MatPaginator;
+  @ViewChild(MatSort) sort!: MatSort;
+
+  ngOnDestroy(): void {
+    this.destroy();
+  }
+
+  ngOnInit(): void {
+    setTimeout(() => {
+      this.init(this.sort, this.paginator);
+    });
+  }
+}
